Removed two casts

+++ lib/docker/conformations.cpp
@@ -111,7 +111,7 @@ namespace Docker {
 			__rmsd_sq.init(confs.size(), confs.size());
 			for (size_t i = 0; i < confs.size(); ++i) {
 				for (size_t j = i + 1; j < confs.size(); ++j) {
-					__rmsd_sq.data[i][j] = __rmsd_sq.data[j][i] = (float) Geom3D::compute_rmsd_sq(confs[i], confs[j]);
+					__rmsd_sq.data[i][j] = __rmsd_sq.data[j][i] = Geom3D::compute_rmsd_sq(confs[i], confs[j]);
 				}
 			}
 			dbgmsg("after rmsd calculation");
